Fix +/- mismatch between player profile and league leaders, fix rng crash in run-meters tracking
- seasonStatsForPlayer was coalescing league-only and league+cup stats per field independently (always taking the higher value), so derived stats like +/- didn't match either scope. Now picks one consistent source, preferring league-only to match the Leaders board.
- Pass leaguePlayerStats through to PlayerProfileModal from ScoutingCenterView/TransfersView/TeamProfileView so those views use the same scope too.
- point.js's per-point run-distance estimate called rng() as a function, but rng is an object with .float() everywhere else in this file - was throwing on every simulated point and silently zeroing out throw/catch meters and completion% for the match.

+++ src/matchEngine/point.js
@@ -73,7 +73,7 @@ function playerLabel(p) {
  * WSZYSTKICH meczów ligowych (nie tylko granych przez gracza pełnym silnikiem).
  */
 function estimateRunMetersForPoint(rng, throwCount = 0) {
-  const base = 90 + rng() * 60
+  const base = 90 + rng.float() * 60
   const extra = Math.min(60, throwCount * 6)
   return base + extra
 }
